Window.lpp {spatstat.linnet}	R Documentation

Extract Window of Spatial Object on a Network

Description

Given a spatial object on a network, these functions extract the window in which the network is defined.

Usage

 ## S3 method for class 'lpp'
Window(X, ...)

 ## S3 method for class 'lppm'
Window(X, ...)

Arguments

X
A spatial object.

...
Ignored.

Details

These are methods for the generic function Window which extract the spatial window in which the object X is defined.

For the methods defined here, X should be a spatial object on a linear network (object of class "lpp" or "lppm").

Value

An object of class "owin" (see owin.object) specifying an observation window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

Window.

Examples

   Window(spiders)

[Package spatstat.linnet version 3.2-5 Index]