as.data.frame.lintess {spatstat.linnet}	R Documentation

Convert Network Tessellation to Data Frame

Description

Converts a tessellation on a linear network into a data frame.

Usage

## S3 method for class 'lintess'
as.data.frame(x, ...)

Arguments

x
Tessellation on a linear network (object of class "lintess").

...
Further arguments passed to as.data.frame.default to determine the row names and other features.

Details

A tessellation on a linear network is a partition of the network into non-overlapping pieces (tiles). Each tile consists of one or more line segments which are subsets of the line segments making up the network. A tile can consist of several disjoint pieces.

This function converts the tessellation x to a data frame. Each row of the data frame specifies one sub-segment of the network, and allocates it to a particular tile. The data frame has the following columns:

The seg column specifies which line segment of the network contains the sub-segment. Values of seg are integer indices for the network segments in as.psp(as.linnet(x)).

The t0 and t1 columns specify the start and end points of the sub-segment. They are numeric values between 0 and 1 inclusive, where the values 0 and 1 representing the network vertices that are joined by this network segment.

The tile column specifies which tile of the tessellation includes this sub-segment. It is a factor whose levels are the names of the tiles.

The tessellation may have marks, which are attached to the tiles of the tessellation. If marks are present, the resulting data frame includes columns containing, for each sub-segment, the mark value of the corresponding tile.

Value

A data frame with columns named seg, t0, t1, tile, and possibly other columns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lintess

Examples

  X <- lineardirichlet(runiflpp(3, simplenet))
  marks(X) <- letters[1:3]
  as.data.frame(X)

[Package spatstat.linnet version 3.2-5 Index]