as.linnet.linim {spatstat.linnet}	R Documentation

Extract Linear Network from Data on a Linear Network

Description

Given some kind of data on a linear network, the command as.linnet extracts the linear network itself.

Usage

 ## S3 method for class 'linim'
as.linnet(X, ...)

 ## S3 method for class 'linfun'
as.linnet(X, ...)

 ## S3 method for class 'lintess'
as.linnet(X, ...)

 ## S3 method for class 'lpp'
as.linnet(X, ..., fatal=TRUE, sparse)

Arguments

X
Data on a linear network. A point pattern (class "lpp"), pixel image (class "linim"), function (class "linfun") or tessellation (class "lintess") on a linear network.

...
Ignored.

fatal
Logical value indicating whether data in the wrong format should lead to an error (fatal=TRUE) or a warning (fatal=FALSE).

sparse
Logical value indicating whether to use a sparse matrix representation, as explained in linnet. Default is to keep the same representation as in X.

Details

These are methods for the generic as.linnet for various classes.

The network on which the data are defined is extracted.

Value

A linear network (object of class "linnet").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

linnet, methods.linnet.

Examples

  # make some data
  xcoord <- linfun(function(x,y,seg,tp) { x }, simplenet)
  as.linnet(xcoord)
  X <- as.linim(xcoord)
  as.linnet(X)

[Package spatstat.linnet version 3.2-5 Index]