as.owin.lpp {spatstat.linnet}	R Documentation

Convert Data on a Network to class owin

Description

Converts data on a linear network into an object of class "owin".

Usage

 ## S3 method for class 'lpp'
as.owin(W, ..., fatal=TRUE)

 ## S3 method for class 'lppm'
as.owin(W, ..., fatal=TRUE)

Arguments

W
Data specifying an observation window, in any of several formats described under Details below.

fatal
Logical value determining what to do if the data cannot be converted to an observation window. See Details.

...
Ignored.

Details

The class "owin" is a way of specifying the observation window for a point pattern. See owin.object for an overview.

The function as.owin converts data in any of several formats into an object of class "owin" for use by the spatstat package. The function as.owin is generic, with methods for different classes of objects, and a default method.

A long list of methods for as.owin is documented in the help file for as.owin in the spatstat.geom package.

This help file documents additional methods applicable when W is

an object of class "lpp" representing a point pattern on a linear network. In this case, as.owin extracts the linear network and returns a window containing this network.

an object of class "lppm" representing a fitted point process model on a linear network. In this case, as.owin extracts the linear network and returns a window containing this network.

If the argument W cannot be converted to a window, then an error will be generated (if fatal=TRUE) or a value of NULL will be returned (if fatal=FALSE).

Value

An object of class "owin" (see owin.object) specifying an observation window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

as.owin, owin.object, owin.

Additional methods for as.owin may be provided by other packages outside the spatstat family.

Examples

  as.owin(simplenet)

[Package spatstat.linnet version 3.2-5 Index]