begins {spatstat.linnet}	R Documentation

Check Start of Character String

Description

Checks whether a character string begins with a particular prefix.

Usage

begins(x, firstbit)

Arguments

x
Character string, or vector of character strings, to be tested.

firstbit
A single character string.

Details

This simple wrapper function checks whether (each entry in) x begins with the string firstbit, and returns a logical value or logical vector with one entry for each entry of x. This function is useful mainly for reducing complexity in model formulae.

Value

Logical vector of the same length as x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

Examples

  begins(c("Hello", "Goodbye"), "Hell")
  begins("anything", "")

[Package spatstat.linnet version 3.2-5 Index]