bw.voronoi {spatstat.linnet}	R Documentation

Cross Validated Bandwidth Selection for Voronoi Estimator of Intensity on a Network

Description

Uses cross-validation to select a smoothing bandwidth for the Voronoi estimate of point process intensity on a linear network.

Usage

bw.voronoi(X, ..., probrange = c(0.2, 0.8), nprob = 10,
           prob = NULL, nrep = 100, verbose = TRUE, warn=TRUE)

Arguments

X
Point pattern on a linear network (object of class "lpp").

...
Ignored.

probrange
Numeric vector of length 2 giving the range of bandwidths (retention probabilities) to be assessed.

nprob
Integer. Number of bandwidths to be assessed.

prob
Optional. A numeric vector of bandwidths (retention probabilities) to be assessed. Entries must be probabilities between 0 and 1. Overrides nprob and probrange.

nrep
Number of simulated realisations to be used for the computation.

verbose
Logical value indicating whether to print progress reports.

warn
Logical. If TRUE, issue a warning if the maximum of the cross-validation criterion occurs at one of the ends of the search interval.

Details

This function uses likelihood cross-validation to choose the optimal value of the thinning fraction f (the retention probability) to be used in the smoothed Voronoi estimator of point process intensity densityVoronoi.lpp.

Value

A single numerical value giving the selected bandwidth. The result also belongs to the class "bw.optim" (see bw.optim.object) which can be plotted to show the bandwidth selection criterion as a function of sigma.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk and Mehdi Moradi m2.moradi@yahoo.com.

References

Moradi, M., Cronie, 0., Rubak, E., Lachieze-Rey, R., Mateu, J. and Baddeley, A. (2019) Resample-smoothing of Voronoi intensity estimators. Statistics and Computing 29 (5) 995–1010.

See Also

densityVoronoi.lpp, bw.optim.object

Examples

   np <- if(interactive()) 10 else 3
   nr <- if(interactive()) 100 else 2
   b <- bw.voronoi(spiders, nprob=np, nrep=nr)
   b
   plot(b)

[Package spatstat.linnet version 3.2-5 Index]