chop.linnet {spatstat.linnet}	R Documentation

Divide a Linear Network into Tiles Using Infinite Lines

Description

Given a linear network and a set of infinite lines, divide the network into tiles demarcated by the lines. The result is a tessellation of the network.

Usage

chop.linnet(X, L)

Arguments

X
Linear network (object of class "linnet") or data acceptable to as.linnet.

L
Infinite line or lines (object of class "infline").

Details

The first line of L divides X into two tiles. Subsequent lines divide each of these tiles. The result is a tessellation of X. Tiles are not necessarily connected sets.

Value

Tessellation on a linear network (object of class "lintess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

crossing.linnet to determine the crossing points between the lines and the network.

divide.linnet to divide a network into a tessellation using arbitrary cut points.

Examples

   L <- infline(p=runif(3), theta=runif(3, max=pi/2))
   Y <- chop.linnet(simplenet, L)
   plot(Y, main="")
   plot(L, col="red")

[Package spatstat.linnet version 3.2-5 Index]