crossing.linnet {spatstat.linnet}	R Documentation

Crossing Points between Linear Network and Other Lines

Description

Find all the crossing-points between a linear network and another pattern of lines or line segments.

Usage

crossing.linnet(X, Y)

Arguments

X
Linear network (object of class "linnet").

Y
A linear network, or a spatial pattern of line segments (class "psp") or infinite lines (class "infline").

Details

All crossing-points between X and Y are determined. The result is a point pattern on the network X.

Value

Point pattern on a linear network (object of class "lpp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

crossing.psp

Examples

   plot(simplenet, main="")
   L <- infline(p=runif(3), theta=runif(3, max=pi/2))
   plot(L, col="red")
   Y <- crossing.linnet(simplenet, L)
   plot(Y, add=TRUE, cols="blue")

[Package spatstat.linnet version 3.2-5 Index]