data.lppm {spatstat.linnet}	R Documentation

Extract Original Data from a Fitted Point Process Model on a Network

Description

Given a fitted point process model on a linear network, this function extracts the original point pattern dataset to which the model was fitted.

Usage

  data.lppm(object)

Arguments

object
fitted point process model on a linear network (an object of class "lppm").

Details

An object of class "lppm" represents a point process model that has been fitted to a point pattern dataset on a linear network. It is typically produced by the model-fitting algorithm lppm. The object contains complete information about the original data point pattern to which the model was fitted. This function extracts the original data pattern.

Value

A point pattern on a linear network (object of class "lpp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lppm, data.ppm

Examples

 fit <- lppm(spiders ~ x)
 X <- data.lppm(fit)
 # 'X' is identical to 'spiders'

[Package spatstat.linnet version 3.2-5 Index]