delaunayNetwork {spatstat.linnet}	R Documentation

Linear Network of Delaunay Triangulation or Dirichlet Tessellation

Description

Computes the edges of the Delaunay triangulation or Dirichlet tessellation of a point pattern, and returns the result as a linear network object.

Usage

delaunayNetwork(X)

dirichletNetwork(X, ...)

Arguments

X
A point pattern (object of class "ppp").

...
Arguments passed to as.linnet.psp

Details

For delaunayNetwork, points of X which are neighbours in the Delaunay triangulation (see delaunay) will be joined by a straight line. The result will be returned as a linear network (object of class "linnet").

For dirichletNetwork, the Dirichlet tessellation is computed (see dirichlet) and the edges of the tiles of the tessellation are extracted. This is converted to a linear network using as.linnet.psp.

Value

Linear network (object of class "linnet") or NULL.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

delaunay, dirichlet, delaunayDistance

Examples

  LE <- delaunayNetwork(cells)
  LI <- dirichletNetwork(cells)

[Package spatstat.linnet version 3.2-5 Index]