density.linnet {spatstat.linnet}	R Documentation

Kernel Smoothing of Linear Network

Description

Compute a kernel smoothed intensity function for the line segments of a linear network.

Usage

  ## S3 method for class 'linnet'
density(x, ...)

Arguments

x
Linear network (object of class "linnet")

...
Arguments passed to density.psp to control the amount of smoothing and the spatial resolution of the result.

Details

This is the method for the generic function density for the class "linnet" (linear networks).

The network x is first converted to a line segment pattern (object of class "psp"). Then the method density.psp is applied to the segment pattern.

A kernel estimate of the intensity of the line segment pattern is computed. The result is the convolution of the isotropic Gaussian kernel, of standard deviation sigma, with the line segments.

The intensity of a line segment pattern is the (spatially-varying) amount of segment length per unit area, expressed in the same units as the coordinates of x. If the units of x are in metres, then an intensity value of 3 means that there are 3 metres of segment length per square metre of spatial domain.

See density.psp for more details.

Value

A pixel image in two dimensions (object of class "im") or a numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

density.psp, im.object, density.

Examples

  D <- density(simplenet, 0.1)
  plot(D)
  plot(simplenet, add=TRUE, col="white")
  ## compare with average intensity
  volume(simplenet)/area(Window(simplenet))

[Package spatstat.linnet version 3.2-5 Index]