densityVoronoi.lpp {spatstat.linnet}	R Documentation

Intensity Estimate of Point Pattern on Linear Network Using Voronoi-Dirichlet Tessellation

Description

Computes an adaptive estimate of the intensity function of a point pattern on a linear network, using the Dirichlet-Voronoi tessellation on the network.

Usage

## S3 method for class 'lpp'
densityVoronoi(X, f = 1, ..., nrep = 1, verbose = TRUE)

Arguments

X
Point pattern on a linear network (object of class "lpp").

f
Fraction (between 0 and 1 inclusive) of the data points that will be used to build a tessellation for the intensity estimate.

...
Arguments passed to linim determining the pixel resolution of the result.

nrep
Number of independent repetitions of the randomised procedure.

verbose
Logical value indicating whether to print progress reports.

Details

This function is an alternative to density.lpp. It computes an estimate of the intensity function of a point pattern dataset on a linear network. The result is a pixel image on the network, giving the estimated intensity.

This function is a method for the generic densityVoronoi for the class "lpp" of point patterns on a linear network.

If f=1 (the default), the Voronoi estimate (Barr and Schoenberg, 2010) is computed: the point pattern X is used to construct a Voronoi/Dirichlet tessellation on the network (see lineardirichlet); the lengths of the Dirichlet tiles are computed; the estimated intensity in each tile is the reciprocal of the tile length. The result is a pixel image of intensity estimates which are constant on each tile of the tessellation.

If f=0, the intensity estimate at every location is equal to the average intensity (number of points divided by network length). The result is a pixel image of intensity estimates which are constant.

If f is strictly between 0 and 1, the smoothed Voronoi estimate (Moradi et al, 2019) is computed. The dataset X is randomly thinned by deleting or retaining each point independently, with probability f of retaining a point. The thinned pattern is used to construct a Dirichlet tessellation and form the Voronoi estimate, which is then adjusted by a factor 1/f. This procedure is repeated nrep times and the results are averaged to obtain the smoothed Voronoi estimate.

The value f can be chosen automatically by bandwidth selection using bw.voronoi.

Value

Pixel image on a linear network (object of class "linim").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk and Mehdi Moradi m2.moradi@yahoo.com.

References

Moradi, M., Cronie, 0., Rubak, E., Lachieze-Rey, R., Mateu, J. and Baddeley, A. (2019) Resample-smoothing of Voronoi intensity estimators. Statistics and Computing 29 (5) 995–1010.

See Also

densityVoronoi is the generic, with a method for class "ppp".

lineardirichlet computes the Dirichlet-Voronoi tessellation on a network.

bw.voronoi performs bandwidth selection of the fraction f.

See also density.lpp.

Examples

   nr <- if(interactive()) 100 else 3
   plot(densityVoronoi(spiders, 0.1, nrep=nr))

[Package spatstat.linnet version 3.2-5 Index]