diameter.linnet {spatstat.linnet}	R Documentation

Diameter and Bounding Radius of a Linear Network

Description

Compute the diameter or bounding radius of a linear network measured using the shortest path distance.

Usage

## S3 method for class 'linnet'
diameter(x)

## S3 method for class 'linnet'
boundingradius(x, ...)

Arguments

x
Linear network (object of class "linnet").

...
Ignored.

Details

The diameter of a linear network (in the shortest path distance) is the maximum value of the shortest-path distance between any two points u and v on the network.

The bounding radius of a linear network (in the shortest path distance) is the minimum value, over all points u on the network, of the maximum shortest-path distance from u to another point v on the network.

The functions boundingradius and diameter are generic; the functions boundingradius.linnet and diameter.linnet are the methods for objects of class linnet.

Value

A single numeric value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

boundingradius, diameter, linnet

Examples

   diameter(simplenet)
   boundingradius(simplenet)

[Package spatstat.linnet version 3.2-5 Index]