divide.linnet {spatstat.linnet}	R Documentation

Divide Linear Network at Cut Points

Description

Make a tessellation of a linear network by dividing it into pieces demarcated by the points of a point pattern.

Usage

 divide.linnet(X)

Arguments

X
Point pattern on a linear network (object of class "lpp").

Details

The points X are interpreted as dividing the linear network L=as.linnet(X) into separate pieces.

Two locations on L belong to the same piece if and only if they can be joined by a path in L that does not cross any of the points of X.

The result is a tessellation of the network (object of class "lintess") representing the division of L into pieces.

Value

A tessellation on a linear network (object of class "lintess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk and Greg McSwiggan.

See Also

linnet, lintess.

Examples

  X <- runiflpp(5, simplenet)
  plot(divide.linnet(X))
  plot(X, add=TRUE, pch=16, show.network=FALSE)

[Package spatstat.linnet version 3.2-5 Index]