domain.lpp {spatstat.linnet}	R Documentation

Extract the Linear Network on which Spatial Data are Defined

Description

Given a spatial object representing data on a linear network, extract the network.

Usage

 ## S3 method for class 'lpp'
domain(X, ...)

 ## S3 method for class 'lppm'
domain(X, ...)

 ## S3 method for class 'linfun'
domain(X, ...)

 ## S3 method for class 'lintess'
domain(X, ...)

Arguments

X
A spatial object representing data on a linear network. An object of class "lpp", "lppm", "linfun" or "lintess".

...
Extra arguments. They are ignored by all the methods listed here.

Details

The function domain is generic, with methods for many classes.

For a spatial object X domain(X) extracts the spatial domain in which X is defined.

For a two-dimensional object X, typically domain(X) is the same as Window(X).

The exception is that, if X is a point pattern on a linear network (class "lpp") or a point process model on a linear network (class "lppm"), then domain(X) is the linear network on which the points lie, while Window(X) is the two-dimensional window containing the linear network.

Value

A linear network (object of class "linnet").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

domain, domain.rmhmodel, domain.ppm.

Window, Frame

Examples

  domain(chicago)

[Package spatstat.linnet version 3.2-5 Index]