fitted.lppm {spatstat.linnet}	R Documentation

Fitted Intensity for Point Process on Linear Network

Description

Given a point process model fitted to a point pattern on a linear network, compute the fitted intensity of the model at the points of the pattern, or at the points of the quadrature scheme used to fit the model.

Usage

## S3 method for class 'lppm'
fitted(object, ...,
                      dataonly = FALSE, new.coef = NULL,
		      leaveoneout = FALSE)

Arguments

object
Fitted point process model on a linear network (object of class "lppm").

...
Ignored.

dataonly
Logical value indicating whether to computed fitted intensities at the points of the original point pattern dataset (dataonly=TRUE) or at all the quadrature points of the quadrature scheme used to fit the model (dataonly=FALSE, the default).

new.coef
Numeric vector of parameter values to replace the fitted model parameters coef(object).

leaveoneout
Logical. If TRUE the fitted value at each data point will be computed using a leave-one-out method. See Details.

Details

This is a method for the generic function fitted for the class "lppm" of fitted point process models on a linear network.

The locations u at which the fitted conditional intensity/trend is evaluated, are the points of the quadrature scheme used to fit the model in ppm. They include the data points (the points of the original point pattern dataset x) and other “dummy” points in the window of observation.

If leaveoneout=TRUE, fitted values will be computed for the data points only, using a ‘leave-one-out’ rule: the fitted value at X[i] is effectively computed by deleting this point from the data and re-fitting the model to the reduced pattern X[-i], then predicting the value at X[i]. (Instead of literally performing this calculation, we apply a Taylor approximation using the influence function computed in dfbetas.ppm.

Value

A vector containing the values of the fitted spatial trend.

Entries in this vector correspond to the quadrature points (data or dummy points) used to fit the model. The quadrature points can be extracted from object by union.quad(quad.ppm(object)).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lppm, predict.lppm

Examples

   fit <- lppm(spiders~x+y)
   a <- fitted(fit)
   b <- fitted(fit, dataonly=TRUE)

[Package spatstat.linnet version 3.2-5 Index]