heatkernelapprox {spatstat.linnet}	R Documentation

Approximation to Heat Kernel on Linear Network at Source Point

Description

Computes an approximation to the value of the heat kernel on a network evaluated at its source location.

Usage

heatkernelapprox(X, sigma, nmax = 20, floored=TRUE)

Arguments

X
Point pattern on a linear network (object of class "lpp").

sigma
Numeric. Bandwidth for kernel.

nmax
Number of terms to be used in the sum.

floored
Logical. If TRUE, all values are constrained to be greater than or equal to 1/L where L is the total length of the network. This the exact value of the heat kernel when the bandwidth is infinite.

Details

For each point X[i] in the pattern X, this algorithm computes an approximation to the value of the heat kernel with source point X[i] evaluated at the same location.

The heat kernel κ(u,v) for a source location u evaluated at location v can be expressed as an infinite sum of contributions from all possible paths from u to v. This algorithm applies to the special case u=v where the source point and the query point are the same.

The algorithm computes an approximation to κ(u,u) by taking only the contributions from paths which (a) remain in the line segment containing the point u and (b) visit a vertex at most nmax times.

Value

Numeric vector with one entry for each point in X.

Author(s)

Greg McSwiggan and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

hotrod

Examples

  X <- runiflpp(3,simplenet)
  heatkernelapprox(X, 0.5)

[Package spatstat.linnet version 3.2-5 Index]