integral.linim {spatstat.linnet}	R Documentation

Integral on a Linear Network

Description

Computes the integral (total value) of a function or pixel image over a linear network.

Usage

## S3 method for class 'linim'
integral(f, domain=NULL, weight=NULL, ...)

## S3 method for class 'linfun'
integral(f, domain=NULL, weight=NULL, ..., delta, nd)

Arguments

f
A pixel image on a linear network (class "linim") or a function on a linear network (class "linfun").

domain
Optional window specifying the domain of integration. Alternatively a tessellation.

weight
Optional numerical weight function for the integration. A pixel image (object of class "linim" or "im"), a function (object of class "linfun", "funxy" or a a function(x,y)) or anything acceptable to as.linim.

...
Ignored.

delta
Optional. The step length (in coordinate units) for computing the approximate integral. A single positive number.

nd
Optional. Integer giving the approximate number of sample points on the network.

Details

The integral (total value of the function over the network) is calculated.

If domain is a window (class "owin") then the integration will be restricted to this window. If domain is a tessellation (class "tess") then the integral of f in each tile of domain will be computed.

If weight is given, effectively the integral of weight * f is computed.

Value

A single numeric or complex value (or a vector of such values if domain is a tessellation).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

linim, integral.im

Examples

  # make some data
  xcoord <- linfun(function(x,y,seg,tp) { x }, simplenet)
  integral(xcoord)
  X <- as.linim(xcoord)
  integral(X)

  # integrals inside each tile of a tessellation
  A <- quadrats(Frame(simplenet), 3)
  integral(X, A)

[Package spatstat.linnet version 3.2-5 Index]