intensity.lpp {spatstat.linnet}	R Documentation

Empirical Intensity of Point Pattern on Linear Network

Description

Computes the average number of points per unit length in a point pattern on a linear network.

Usage

## S3 method for class 'lpp'
intensity(X, ...)

Arguments

X
A point pattern on a linear network (object of class "lpp").

...
Ignored.

Details

This is a method for the generic function intensity It computes the empirical intensity of a point pattern on a linear network (object of class "lpp"), i.e. the average density of points per unit length.

If the point pattern is multitype, the intensities of the different types are computed separately.

Value

A numeric value (giving the intensity) or numeric vector (giving the intensity for each possible type).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

intensity, intensity.ppp

Examples

  intensity(chicago)

[Package spatstat.linnet version 3.2-5 Index]