intersect.lintess {spatstat.linnet}	R Documentation

Intersection of Tessellations on a Linear Network

Description

Yields the intersection (common refinement) of two tessellations on a linear network.

Usage

intersect.lintess(X, Y)

Arguments

X,Y
Tessellations (objects of class "lintess") on the same linear network, or data that define such tessellations. See Details.

Details

X and Y should be tessellations on a linear network (objects of class "lintess") and should be defined on the same network. The algorithm finds the common refinement of the two tessellations. Each tile in the resulting tessellation is the intersection of a tile of X with a tile of Y.

Alternatively, one of the arguments X or Y can be a two-dimensional tessellation (object of class "tess") while the other argument is a network or a tessellation on a network. The two-dimensional tessellation will be intersected with the network to produce a tessellation on the network, then intersected with the other tessellation on the network.

Value

Another tessellation (object of class "lintess") on the same linear network as X and Y.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

lintess, divide.linnet, chop.linnet

Examples

  X <- divide.linnet(runiflpp(4, simplenet))
  Y <- divide.linnet(runiflpp(3, simplenet))
  opa <- par(mfrow=c(1,3))
  plot(X)
  plot(Y)
  plot(intersect.lintess(X,Y))
  par(opa)

[Package spatstat.linnet version 3.2-5 Index]