is.marked.lppm {spatstat.linnet}	R Documentation

Test Whether A Point Process Model is Marked

Description

Tests whether a fitted point process model on a network involves “marks” attached to the points.

Usage

  ## S3 method for class 'lppm'
is.marked(X, ...) 

Arguments

X
Fitted point process model on a linear networ (object of class "lppm") usually obtained from lppm.

...
Ignored.

Details

“Marks” are observations attached to each point of a point pattern. For example the chicago dataset contains the locations of crimes, each crime location being marked by the type of crime.

The argument X is a fitted point process model on a network (an object of class "lppm") typically obtained by fitting a model to point pattern data using lppm.

This function returns TRUE if the original data (to which the model X was fitted) were a marked point pattern.

Note that this is not the same as testing whether the model involves terms that depend on the marks (i.e. whether the fitted model ignores the marks in the data). See the Examples for a trick to do this.

If this function returns TRUE, the implications are (for example) that any simulation of this model will require simulation of random marks as well as random point locations.

Value

Logical value, equal to TRUE if X is a model that was fitted to a marked point pattern dataset.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

is.marked.

Examples

  fit <- lppm(chicago ~ x)
  is.marked(fit)
  ## result is TRUE, i.e. the data are marked

  ## To check whether the model involves marks:
  "marks" %in% spatstat.utils::variablesinformula(formula(fit))

[Package spatstat.linnet version 3.2-5 Index]