is.stationary.lppm {spatstat.linnet}	R Documentation

Recognise Stationary and Poisson Point Process Models on a Network

Description

Given a point process model that has been fitted to data on a network, determine whether the model is a stationary point process, and whether it is a Poisson point process.

Usage

## S3 method for class 'lppm'
is.stationary(x)

## S3 method for class 'lppm'
is.poisson(x)

Arguments

x
A fitted spatial point process model on a linear network (object of class "lppm").

Details

The argument x represents a fitted spatial point process model on a linear network.

is.stationary(x) returns TRUE if x represents a stationary point process, and FALSE if not.

is.poisson(x) returns TRUE if x represents a Poisson point process, and FALSE if not.

The functions is.stationary and is.poisson are generic, with methods for many classes of models.

Value

A logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

is.marked to determine whether a model is a marked point process.

is.stationary, is.poisson for generics.

summary.lppm for detailed information.

Model-fitting function lppm.

Examples

  fit <- lppm(spiders ~ x)
  is.stationary(fit)
  is.poisson(fit)

[Package spatstat.linnet version 3.2-5 Index]