linearK {spatstat.linnet}	R Documentation

Linear K Function

Description

Computes an estimate of the linear K function for a point pattern on a linear network.

Usage

linearK(X, r=NULL, ..., correction="Ang", ratio=FALSE)

Arguments

X
Point pattern on linear network (object of class "lpp").

r
Optional. Numeric vector of values of the function argument r. There is a sensible default.

...
Ignored.

correction
Geometry correction. Either "none" or "Ang". See Details.

ratio
Logical. If TRUE, the numerator and denominator of the estimate will also be saved, for use in analysing replicated point patterns.

Details

This command computes the linear K function from point pattern data on a linear network.

If correction="none", the calculations do not include any correction for the geometry of the linear network. The result is the network K function as defined by Okabe and Yamada (2001).

If correction="Ang", the pair counts are weighted using Ang's correction (Ang, 2010; Ang et al, 2012).

Value

Function value table (object of class "fv").

Author(s)

Ang Qi Wei aqw07398@hotmail.com and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Ang, Q.W. (2010) Statistical methodology for spatial point patterns on a linear network. MSc thesis, University of Western Australia.

Ang, Q.W., Baddeley, A. and Nair, G. (2012) Geometrically corrected second-order analysis of events on a linear network, with applications to ecology and criminology. Scandinavian Journal of Statistics 39, 591–617.

Okabe, A. and Yamada, I. (2001) The K-function method on a network and its computational implementation. Geographical Analysis 33, 271-290.

See Also

compileK, lpp

Examples

  X <- rpoislpp(5, simplenet)
  linearK(X)
  linearK(X, correction="none")

[Package spatstat.linnet version 3.2-5 Index]