linearKcross {spatstat.linnet}	R Documentation

Multitype K Function (Cross-type) for Linear Point Pattern

Description

For a multitype point pattern on a linear network, estimate the multitype K function which counts the expected number of points of type j within a given distance of a point of type i.

Usage

linearKcross(X, i, j, r=NULL, ..., correction="Ang")

Arguments

X
The observed point pattern, from which an estimate of the cross type K function Kij(r) will be computed. An object of class "lpp" which must be a multitype point pattern (a marked point pattern whose marks are a factor).

i
Number or character string identifying the type (mark value) of the points in X from which distances are measured. Defaults to the first level of marks(X).

j
Number or character string identifying the type (mark value) of the points in X to which distances are measured. Defaults to the second level of marks(X).

r
numeric vector. The values of the argument r at which the K-function Kij(r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

correction
Geometry correction. Either "none" or "Ang". See Details.

...
Ignored.

Details

This is a counterpart of the function Kcross for a point pattern on a linear network (object of class "lpp").

The arguments i and j will be interpreted as levels of the factor marks(X). If i and j are missing, they default to the first and second level of the marks factor, respectively.

The argument r is the vector of values for the distance r at which Kij(r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must not exceed the radius of the largest disc contained in the window.

Value

An object of class "fv" (see fv.object).

Warnings

The arguments i and j are interpreted as levels of the factor marks(X). Beware of the usual trap with factors: numerical values are not interpreted in the same way as character values.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Baddeley, A, Jammalamadaka, A. and Nair, G. (2014) Multitype point process analysis of spines on the dendrite network of a neuron. Applied Statistics (Journal of the Royal Statistical Society, Series C), 63, 673–694.

See Also

linearKdot, linearK.

Examples

   K <- linearKcross(chicago, "assault", "robbery")

[Package spatstat.linnet version 3.2-5 Index]