linearKdot.inhom {spatstat.linnet}	R Documentation

Inhomogeneous multitype K Function (Dot-type) for Linear Point Pattern

Description

For a multitype point pattern on a linear network, estimate the inhomogeneous multitype K function which counts the expected number of points (of any type) within a given distance of a point of type i.

Usage

linearKdot.inhom(X, i, lambdaI=NULL, lambdadot=NULL, r=NULL, ...,
                 correction="Ang", normalise=TRUE, sigma=NULL)

Arguments

X
The observed point pattern, from which an estimate of the dot type K function K[i.](r) will be computed. An object of class "lpp" which must be a multitype point pattern (a marked point pattern whose marks are a factor).

i
Number or character string identifying the type (mark value) of the points in X from which distances are measured. Defaults to the first level of marks(X).

lambdaI
Intensity values for the points of type i. Either a numeric vector, a function, a pixel image (object of class "im" or "linim") or a fitted point process model (object of class "ppm" or "lppm") or NULL.

lambdadot
Intensity values for all points of X. Either a numeric vector, a function, a pixel image (object of class "im" or "linim") or a fitted point process model (object of class "ppm" or "lppm") or NULL.

r
numeric vector. The values of the argument r at which the K-function K[i.](r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

correction
Geometry correction. Either "none" or "Ang". See Details.

...
Arguments passed to lambdaI and lambdadot if they are functions.

normalise
Logical. If TRUE (the default), the denominator of the estimator is data-dependent (equal to the sum of the reciprocal intensities at the points of type i), which reduces the sampling variability. If FALSE, the denominator is the length of the network.

sigma
Smoothing bandwidth passed to density.lpp for estimation of intensities when either lambdaI or lambdadot is NULL.

Details

This is a counterpart of the function Kdot.inhom for a point pattern on a linear network (object of class "lpp").

The argument i will be interpreted as levels of the factor marks(X). If i is missing, it defaults to the first level of the marks factor.

The argument r is the vector of values for the distance r at which Ki.(r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must not exceed the radius of the largest disc contained in the window.

If lambdaI or lambdadot is missing, it will be estimated by kernel smoothing using density.lpp.

If lambdaI or lambdadot is a fitted point process model, the default behaviour is to update the model by re-fitting it to the data, before computing the fitted intensity. This can be disabled by setting update=FALSE.

Value

An object of class "fv" (see fv.object).

Warnings

The argument i is interpreted as a level of the factor marks(X). Beware of the usual trap with factors: numerical values are not interpreted in the same way as character values.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Baddeley, A, Jammalamadaka, A. and Nair, G. (2014) Multitype point process analysis of spines on the dendrite network of a neuron. Applied Statistics (Journal of the Royal Statistical Society, Series C), 63, 673–694.

See Also

linearKdot, linearK.

Examples

   lam <- table(marks(chicago))/(summary(chicago)$totlength)
   lamI <- function(x,y,const=lam[["assault"]]){ rep(const, length(x)) }
   lam. <- function(x,y,const=sum(lam)){ rep(const, length(x)) }

   K <- linearKdot.inhom(chicago, "assault", lamI, lam.)

   # using fitted models for the intensity
   # fit <- lppm(chicago ~marks + x)
   # linearKdot.inhom(chicago, "assault", fit, fit)

[Package spatstat.linnet version 3.2-5 Index]