lineardirichlet {spatstat.linnet}	R Documentation

Dirichlet Tessellation on a Linear Network

Description

Given a point pattern on a linear network, compute the Dirichlet (or Voronoi or Thiessen) tessellation induced by the points.

Usage

lineardirichlet(X)

Arguments

X
Point pattern on a linear network (object of class "lpp").

Details

The Dirichlet tessellation induced by a point pattern X on a linear network L is a partition of L into subsets. The subset L[i] associated with the data point X[i] is the part of L lying closer to X[i] than to any other data point X[j], where distance is measured by the shortest path.

Value

A tessellation on a linear network (object of class "lintess").

Missing tiles

If the linear network is not connected, and if one of the connected components contains no data points, then the Dirichlet tessellation is mathematically undefined inside this component. The resulting tessellation object includes a tile with label NA, which contains this component of the network. A plot of the tessellation will not show this tile.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

lintess.

For the Dirichlet tessellation in two-dimensional space, see dirichlet.

Examples

  X <- runiflpp(5, simplenet)
  plot(lineardirichlet(X), lwd=3)
  points(X)

[Package spatstat.linnet version 3.2-5 Index]