linearmarkequal {spatstat.linnet}	R Documentation

Mark Connection Function for Multitype Point Pattern on Linear Network

Description

For a multitype point pattern on a linear network, estimate the mark connection function from points of type i to points of type j.

Usage

linearmarkequal(X, r=NULL, ...)

Arguments

X
The observed point pattern, from which an estimate of the mark connection function p[ij](r) will be computed. An object of class "lpp" which must be a multitype point pattern (a marked point pattern whose marks are a factor).

r
numeric vector. The values of the argument r at which the function p[ij](r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

...
Arguments passed to linearpcfcross and linearpcf.

Details

This is the mark equality function for a point pattern on a linear network (object of class "lpp").

The argument r is the vector of values for the distance r at which p[ij](r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must not exceed the radius of the largest disc contained in the window.

Value

An object of class "fv" (see fv.object).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Baddeley, A, Jammalamadaka, A. and Nair, G. (2014) Multitype point process analysis of spines on the dendrite network of a neuron. Applied Statistics (Journal of the Royal Statistical Society, Series C), 63, 673–694.

See Also

linearpcfcross, linearpcf, linearmarkconnect, markconnect.

Examples

   if(interactive()) {
     X <- chicago
   } else {
     m <- sample(factor(c("A","B")), 20, replace=TRUE)
     X <- runiflpp(20, simplenet) %mark% m
   }
   p <- linearmarkequal(X)

[Package spatstat.linnet version 3.2-5 Index]