linearpcf {spatstat.linnet}	R Documentation

Linear Pair Correlation Function

Description

Computes an estimate of the linear pair correlation function for a point pattern on a linear network.

Usage

linearpcf(X, r=NULL, ..., correction="Ang", ratio=FALSE)

Arguments

X
Point pattern on linear network (object of class "lpp").

r
Optional. Numeric vector of values of the function argument r. There is a sensible default.

...
Arguments passed to density.default to control the smoothing.

correction
Geometry correction. Either "none" or "Ang". See Details.

ratio
Logical. If TRUE, the numerator and denominator of each estimate will also be saved, for use in analysing replicated point patterns.

Details

This command computes the linear pair correlation function from point pattern data on a linear network.

The pair correlation function is estimated from the shortest-path distances between each pair of data points, using the fixed-bandwidth kernel smoother density.default, with a bias correction at each end of the interval of r values. To switch off the bias correction, set endcorrect=FALSE.

The bandwidth for smoothing the pairwise distances is determined by arguments ... passed to density.default, mainly the arguments bw and adjust. The default is to choose the bandwidth by Silverman's rule of thumb bw="nrd0" explained in density.default.

If correction="none", the calculations do not include any correction for the geometry of the linear network. The result is an estimate of the first derivative of the network K function defined by Okabe and Yamada (2001).

If correction="Ang", the pair counts are weighted using Ang's correction (Ang, 2010). The result is an estimate of the pair correlation function in the linear network.

Value

Function value table (object of class "fv").

If ratio=TRUE then the return value also has two attributes called "numerator" and "denominator" which are "fv" objects containing the numerators and denominators of each estimate of g(r).

Author(s)

Ang Qi Wei aqw07398@hotmail.com and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Ang, Q.W. (2010) Statistical methodology for spatial point patterns on a linear network. MSc thesis, University of Western Australia.

Ang, Q.W., Baddeley, A. and Nair, G. (2012) Geometrically corrected second-order analysis of events on a linear network, with applications to ecology and criminology. Scandinavian Journal of Statistics 39, 591–617.

Okabe, A. and Yamada, I. (2001) The K-function method on a network and its computational implementation. Geographical Analysis 33, 271-290.

See Also

linearK, linearpcfinhom, lpp

Examples

  X <- rpoislpp(5, simplenet)
  linearpcf(X)
  linearpcf(X, correction="none")

[Package spatstat.linnet version 3.2-5 Index]