linearpcfEuclid {spatstat.linnet}	R Documentation

Linear Pair Correlation Function Using Euclidean Distance

Description

Computes an estimate of the pair correlation function based on Euclidean distances, for a point pattern on a linear network.

Usage

linearpcfEuclid(X, r = NULL, ...)

Arguments

X
Point pattern on linear network (object of class "lpp").

r
Optional. Numeric vector of values of the function argument r. There is a sensible default.

...
Ignored.

Details

This command computes an estimate of the pair correlation function based on Euclidean distances between the points, as described by Rakshit, Nair and Baddeley (2017).

This is different from the linear pair correlation function based on shortest-path distances, which is computed by linearpcf.

The linear pair correlation function based on Euclidean distances is defined in equation (15) of Rakshit, Nair and Baddeley (2017). The estimate is computed from the point pattern as described in equation (31).

Value

Function value table (object of class "fv").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Rakshit. S., Nair, G. and Baddeley, A. (2017) Second-order analysis of point patterns on a network using any distance metric. Spatial Statistics 22 (1) 129–154.

See Also

linearKEuclid, linearpcfEuclidInhom.

See linearpcf for the corresponding function based on shortest-path distances.

Examples

  X <- rpoislpp(5, simplenet)
  g <- linearpcfEuclid(X)

[Package spatstat.linnet version 3.2-5 Index]