linearpcfEuclidInhom {spatstat.linnet}	R Documentation

Inhomogeneous Linear Pair Correlation Function Based on Euclidean Distances

Description

Computes an estimate of the inhomogeneous pair correlation function based on Euclidean distances, for a point pattern on a linear network.

Usage

   linearpcfEuclidInhom(X, lambda = NULL, r = NULL, ...,
      normalise = TRUE, normpower = 2,
      update = TRUE, leaveoneout = TRUE,
      sigma=NULL, adjust.sigma=1, bw="nrd0", adjust.bw=1)

Arguments

X
Point pattern on linear network (object of class "lpp").

lambda
Intensity values for the point pattern. Either a numeric vector, a function, a pixel image (object of class "im" or "linim") or a fitted point process model (object of class "ppm" or "lppm") or NULL.

r
Optional. Numeric vector of values of the function argument r. There is a sensible default.

...
Ignored.

normalise
Logical. If TRUE (the default), the denominator of the estimator is data-dependent (equal to the sum of the reciprocal intensities at the data points, raised to normpower), which reduces the sampling variability. If FALSE, the denominator is the length of the network.

normpower
Integer (usually either 1 or 2). Normalisation power. See Details.

update
Logical value indicating what to do when lambda is a fitted model (class "lppm" or "ppm"). If update=TRUE (the default), the model will first be refitted to the data X (using update.lppm or update.ppm) before the fitted intensity is computed. If update=FALSE, the fitted intensity of the model will be computed without re-fitting it to X.

leaveoneout
Logical value specifying whether to use a leave-one-out rule when calculating the intensity. See Details.

sigma
Smoothing bandwidth (passed to density.lpp) for kernel density estimation of the intensity when lambda=NULL.

adjust.sigma
Numeric value. sigma will be multiplied by this value.

bw
Smoothing bandwidth (passed to density.default) for one-dimensional kernel smoothing of the pair correlation function. Either a numeric value, or a character string recognised by density.default.

adjust.bw
Numeric value. bw will be multiplied by this value.

Details

This command computes the inhomogeneous version of the pair correlation function based on Euclidean distances, for a point pattern on a linear network.

This is different from the inhomogeneous pair correlation function based on shortest-path distances, which is computed by linearpcfinhom.

The inhomogeneous pair correlation function based on Euclidean distances is defined in equation (30) of Rakshit, Nair and Baddeley (2017). Estimation is performed as described in equation (34) of Rakshit, Nair and Baddeley (2017).

The argument lambda should provide estimated values of the intensity of the point process at each point of X.

If lambda=NULL, the intensity will be estimated by kernel smoothing by calling density.lpp with the smoothing bandwidth sigma, and with any other relevant arguments that might be present in .... A leave-one-out kernel estimate will be computed if leaveoneout=TRUE.

If lambda is given, then it may be a numeric vector (of length equal to the number of points in X), or a function(x,y) that will be evaluated at the points of X to yield numeric values, or a pixel image (object of class "im") or a fitted point process model (object of class "ppm" or "lppm").

If lambda is a fitted point process model, the default behaviour is to update the model by re-fitting it to the data, before computing the fitted intensity. This can be disabled by setting update=FALSE. The intensity at data points will be computed by fitted.lppm or fitted.ppm. A leave-one-out estimate will be computed if leaveoneout=TRUE and update=TRUE.

If normalise=TRUE (the default), then the estimate is multiplied by c^normpower where c = length(L)/sum[i] (1/lambda(x[i])). This rescaling reduces the variability and bias of the estimate in small samples and in cases of very strong inhomogeneity. The default value of normpower is 1 (for consistency with previous versions of spatstat) but the most sensible value is 2, which would correspond to rescaling the lambda values so that sum[i] (1/lambda(x[i])) = area(W).

Value

Function value table (object of class "fv").

Warning

Older versions of linearpcfEuclidInhom interpreted lambda=NULL to mean that the homogeneous function linearpcfEuclid should be computed. This was changed to the current behaviour in version 3.1-0 of spatstat.linnet.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Rakshit. S., Nair, G. and Baddeley, A. (2017) Second-order analysis of point patterns on a network using any distance metric. Spatial Statistics 22 (1) 129–154.

See Also

linearKEuclidInhom, linearpcfEuclid.

See linearpcfinhom for the corresponding function based on shortest-path distances.

Examples

  X <- rpoislpp(5, simplenet)
  fit <- lppm(X ~x)
  g <- linearpcfEuclidInhom(X, lambda=fit)
  plot(g)

[Package spatstat.linnet version 3.2-5 Index]