linearpcfdot {spatstat.linnet}	R Documentation

Multitype Pair Correlation Function (Dot-type) for Linear Point Pattern

Description

For a multitype point pattern on a linear network, estimate the multitype pair correlation function from points of type i to points of any type.

Usage

linearpcfdot(X, i, r=NULL, ..., correction="Ang")

Arguments

X
The observed point pattern, from which an estimate of the i-to-any pair correlation function g[i.](r) will be computed. An object of class "lpp" which must be a multitype point pattern (a marked point pattern whose marks are a factor).

i
Number or character string identifying the type (mark value) of the points in X from which distances are measured. Defaults to the first level of marks(X).

r
numeric vector. The values of the argument r at which the function g[i.](r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

correction
Geometry correction. Either "none" or "Ang". See Details.

...
Arguments passed to density.default to control the kernel smoothing.

Details

This is a counterpart of the function pcfdot for a point pattern on a linear network (object of class "lpp").

The argument i will be interpreted as levels of the factor marks(X). If i is missing, it defaults to the first level of the marks factor.

The argument r is the vector of values for the distance r at which g[i.](r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must not exceed the radius of the largest disc contained in the window.

Value

An object of class "fv" (see fv.object).

Warnings

The argument i is interpreted as a level of the factor marks(X). Beware of the usual trap with factors: numerical values are not interpreted in the same way as character values.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Baddeley, A, Jammalamadaka, A. and Nair, G. (2014) Multitype point process analysis of spines on the dendrite network of a neuron. Applied Statistics (Journal of the Royal Statistical Society, Series C), 63, 673–694.

See Also

linearpcfcross, linearpcf, pcfdot.

Examples

   g <- linearpcfdot(chicago, "assault")

[Package spatstat.linnet version 3.2-5 Index]