linequad {spatstat.linnet}	R Documentation

Quadrature Scheme on a Linear Network

Description

Generates a quadrature scheme (an object of class "quad") on a linear network.

Usage

linequad(X, Y, ..., eps = NULL, nd = 1000, random = FALSE)

Arguments

X
Data points. An object of class "lpp" or "ppp".

Y
Line segments on which the points of X lie. An object of class "psp". Required only when X is a "ppp" object.

...
Ignored.

eps
Optional. Spacing between successive dummy points along each segment. (This is the maximum spacing; some spacings will be shorter.)

nd
Optional. Total number of dummy points to be generated. (Actual number may be larger.)

random
Logical value indicating whether the sequence of dummy points should start at a randomly-chosen position along each segment.

Details

This command generates a quadrature scheme (object of class "quad") from a pattern of points on a linear network.

Normally the user does not need to call linequad explicitly. It is invoked by spatstat functions when needed. A quadrature scheme is required by lppm in order to fit point process models to point pattern data on a linear network. A quadrature scheme is also used by rhohat.lpp and other functions.

In order to create the quadrature scheme, dummy points are placed along each line segment of the network. The dummy points are evenly-spaced with spacing eps. The default is eps = totlen/nd where totlen is the total length of all line segments in the network.

Every line segment of the network will contain at least one dummy point. Consequently the actual number of dummy points generated will typically be greater than nd, especially when nd is small. If eps is specified, the number of dummy points will be greater than totlen/eps, especially when eps is large.

Value

A quadrature scheme (object of class "quad").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Greg McSwiggan and Suman Rakshit.

See Also

lppm

[Package spatstat.linnet version 3.2-5 Index]