lppm {spatstat.linnet}	R Documentation

Fit Point Process Model to Point Pattern on Linear Network

Description

Fit a point process model to a point pattern dataset on a linear network

Usage

lppm(X, ...)

## S3 method for class 'formula'
lppm(X, interaction=NULL, ..., data=NULL)

## S3 method for class 'lpp'
lppm(X, ..., eps=NULL, nd=1000, random=FALSE)

Arguments

X
Either an object of class "lpp" specifying a point pattern on a linear network, or a formula specifying the point process model.

...
Arguments passed to ppm.

interaction
An object of class "interact" describing the point process interaction structure, or NULL indicating that a Poisson process (stationary or nonstationary) should be fitted.

data
Optional. The values of spatial covariates (other than the Cartesian coordinates) required by the model. A list whose entries are images, functions, windows, tessellations or single numbers.

eps
Optional. Spacing between dummy points along each segment of the network.

nd
Optional. Total number of dummy points placed on the network. Ignored if eps is given.

random
Logical value indicating whether the grid of dummy points should be placed at a randomised starting position.

Details

This function fits a point process model to data that specify a point pattern on a linear network. It is a counterpart of the model-fitting function ppm designed to work with objects of class "lpp" instead of "ppp".

The function lppm is generic, with methods for the classes formula and lppp.

In lppm.lpp the first argument X should be an object of class "lpp" (created by the command lpp) specifying a point pattern on a linear network.

In lppm.formula, the first argument is a formula in the R language describing the spatial trend model to be fitted. It has the general form pattern ~ trend where the left hand side pattern is usually the name of a point pattern on a linear network (object of class "lpp") to which the model should be fitted, or an expression which evaluates to such a point pattern; and the right hand side trend is an expression specifying the spatial trend of the model.

Other arguments ... are passed from lppm.formula to lppm.lpp and from lppm.lpp to ppm.

Value

An object of class "lppm" representing the fitted model. There are methods for print, predict, coef and similar functions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Greg McSwiggan.

References

Ang, Q.W. (2010) Statistical methodology for events on a network. Master's thesis, School of Mathematics and Statistics, University of Western Australia.

Ang, Q.W., Baddeley, A. and Nair, G. (2012) Geometrically corrected second-order analysis of events on a linear network, with applications to ecology and criminology. Scandinavian Journal of Statistics 39, 591–617.

McSwiggan, G., Nair, M.G. and Baddeley, A. (2012) Fitting Poisson point process models to events on a linear network. Manuscript in preparation.

See Also

methods.lppm, predict.lppm, ppm, lpp.

Examples

  X <- runiflpp(15, simplenet)
  lppm(X ~1)
  lppm(X ~x)
  marks(X) <- factor(rep(letters[1:3], 5))
  lppm(X ~ marks)
  lppm(X ~ marks * x)

[Package spatstat.linnet version 3.2-5 Index]