mean.linim {spatstat.linnet}	R Documentation

Mean, Median, Quantiles of Pixel Values on a Linear Network

Description

Calculates the mean, median, or quantiles of the pixel values in a pixel image on a linear network.

Usage

  ## S3 method for class 'linim'
mean(x, ...)

  ## S3 method for class 'linim'
median(x, ...)

  ## S3 method for class 'linim'
quantile(x, probs=seq(0,1,0.25), ...)

  ## S3 method for class 'linim'
quantilefun(x, ..., type=1)

Arguments

x
A pixel image on a linear network (object of class "linim").

probs
Vector of probabilities for which quantiles should be calculated.

...
Arguments passed to other methods.

type
Integer specifying the type of quantiles, as explained in quantile.default. Only types 1 and 2 are currently implemented.

Details

These functions calculate the mean, median and quantiles of the pixel values in the image x on a linear network.

An object of class "linim" describes a pixel image on a linear network. See linim.

The functions described here are methods for the generic mean, median and quantile for the class "linim".

Value

For mean and median, a single number. For quantile, a numeric vector of the same length as probs. For quantilefun, a function.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

mean, median, quantile,

mean.im, quantile.im, quantilefun

Examples

  M <- psp2mask(as.psp(simplenet))
  Z <- as.im(function(x,y) {x-y}, W=M)
  X <- linim(simplenet, Z)
  X
  mean(X)
  median(X)
  quantile(X)
  f <- quantilefun(X)

[Package spatstat.linnet version 3.2-5 Index]