methods.linnet {spatstat.linnet}	R Documentation

Methods for Linear Networks

Description

These are methods for the class "linnet" of linear networks.

Usage

as.linnet(X, ...)

## S3 method for class 'linnet'
as.linnet(X, ..., sparse, maxsize=30000)

## S3 method for class 'linnet'
as.owin(W, ...)

## S3 method for class 'linnet'
as.psp(x, ..., fatal=TRUE)

## S3 method for class 'linnet'
nsegments(x)

## S3 method for class 'linnet'
nvertices(x, ...)

## S3 method for class 'linnet'
pixellate(x, ...)

## S3 method for class 'linnet'
print(x, ...)

## S3 method for class 'linnet'
summary(object, ...)

## S3 method for class 'linnet'
unitname(x)

## S3 replacement method for class 'linnet'
unitname(x) <- value

vertexdegree(x)

## S3 method for class 'linnet'
vertices(w)

## S3 method for class 'linnet'
volume(x)

## S3 method for class 'linnet'
Window(X, ...)

Arguments

x,X,object,w,W
An object of class "linnet" representing a linear network.

...
Arguments passed to other methods.

value
A valid name for the unit of length for x. See unitname.

fatal
Logical value indicating whether data in the wrong format should lead to an error (fatal=TRUE) or a warning (fatal=FALSE).

sparse
Logical value indicating whether to use a sparse matrix representation, as explained in linnet. Default is to keep the same representation as in X.

maxsize
Maximum permitted number of network vertices (to prevent a system crash due to lack of memory) when creating a network with sparse=FALSE.

Details

The function as.linnet is generic. It converts data from some other format into an object of class "linnet". The method as.linnet.lpp extracts the linear network information from an lpp object. The method as.linnet.linnet converts a linear network into another linear network with the required format.

The other functions are methods for the generic commands as.owin, as.psp, nsegments, nvertices, pixellate, print, summary, unitname, unitname<-, vertices, volume and Window for the class "linnet".

The methods as.owin.linnet and Window.linnet extract the window containing the linear network, and return it as an object of class "owin".

The method as.psp.linnet extracts the lines of the linear network as a line segment pattern (object of class "psp") while nsegments.linnet simply counts the number of line segments.

The method vertices.linnet extracts the vertices (nodes) of the linear network and nvertices.linnet simply counts the vertices. The function vertexdegree calculates the topological degree of each vertex (the number of lines emanating from that vertex) and returns these values as an integer vector.

The method pixellate.linnet applies as.psp.linnet to convert the network to a collection of line segments, then invokes pixellate.psp.

Value

For as.linnet the value is an object of class "linnet". For other functions, see the help file for the corresponding generic function.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

linnet.

Generic functions: as.owin, as.psp, nsegments, nvertices, pixellate, print, summary, unitname, unitname<-, vertices, volume and Window.

Special tools: thinNetwork, insertVertices, joinVertices, connected.linnet.

lixellate for dividing segments into shorter segments.

Examples

  simplenet
  summary(simplenet)
  nsegments(simplenet)
  nvertices(simplenet)
  pixellate(simplenet)
  volume(simplenet)
  unitname(simplenet) <- c("cubit", "cubits")
  Window(simplenet)

[Package spatstat.linnet version 3.2-5 Index]