methods.lpp {spatstat.linnet}	R Documentation

Methods for Point Patterns on a Linear Network

Description

These are methods specifically for the class "lpp" of point patterns on linear networks.

Usage

## S3 method for class 'lpp'
as.ppp(X, ..., fatal=TRUE)

## S3 method for class 'lpp'
as.psp(x, ..., fatal=TRUE)

## S3 replacement method for class 'lpp'
marks(x, ...) <- value

## S3 method for class 'lpp'
nsegments(x)

## S3 method for class 'lpp'
print(x, ...)

## S3 method for class 'summary.lpp'
print(x, ...)

## S3 method for class 'lpp'
summary(object, ...)

## S3 method for class 'lpp'
unitname(x)

## S3 replacement method for class 'lpp'
unitname(x) <- value

## S3 method for class 'lpp'
unmark(X)

Arguments

x,X,object
An object of class "lpp" representing a point pattern on a linear network.

...
Arguments passed to other methods.

value
Replacement value for the marks or unitname of x. See Details.

fatal
Logical value indicating whether data in the wrong format should lead to an error (fatal=TRUE) or a warning (fatal=FALSE).

Details

These are methods for the generic functions as.ppp, as.psp, marks<-, nsegments, print, summary, unitname, unitname<- and unmark for objects of the class "lpp".

For "marks<-.lpp" the replacement value should be either NULL, or a vector of length equal to the number of points in x, or a data frame with one row for each point in x.

For "unitname<-.lpp" the replacement value should be a valid name for the unit of length, as described in unitname.

Value

See the documentation on the corresponding generic function.

Other methods

An object of class "lpp" also inherits the class "ppx" for which many other methods are available. See methods.ppx.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

lpp, intensity.lpp, methods.ppx

Examples

  X <- runiflpp(10, simplenet)
  unitname(X) <- c("furlong", "furlongs")
  X
  summary(X)
  summary(chicago)
  nsegments(X)
  Y <- as.ppp(X)

[Package spatstat.linnet version 3.2-5 Index]