methods.lppm {spatstat.linnet}	R Documentation

Methods for Fitted Point Process Models on a Linear Network

Description

These are methods for the class "lppm" of fitted point process models on a linear network.

Usage

  ## S3 method for class 'lppm'
coef(object, ...)

  ## S3 method for class 'lppm'
emend(object, ...)

  ## S3 method for class 'lppm'
extractAIC(fit, ...)

  ## S3 method for class 'lppm'
formula(x, ...)

  ## S3 method for class 'lppm'
logLik(object, ...)

  ## S3 method for class 'lppm'
deviance(object, ...)

  ## S3 method for class 'lppm'
nobs(object, ...)

  ## S3 method for class 'lppm'
print(x, ...)

  ## S3 method for class 'lppm'
summary(object, ...)

  ## S3 method for class 'lppm'
terms(x, ...)

  ## S3 method for class 'lppm'
update(object, ...)

  ## S3 method for class 'lppm'
valid(object, ...)

  ## S3 method for class 'lppm'
vcov(object, ...)

  ## S3 method for class 'lppm'
as.linnet(X, ...)

  ## S3 method for class 'lppm'
response(object)

Arguments

object,fit,x,X
An object of class "lppm" representing a fitted point process model on a linear network.

...
Arguments passed to other methods, usually the method for the class "ppm".

Details

These are methods for the R generic commands coef, extractAIC, formula, logLik, deviance, nobs, print, summary, terms, update and vcov, and the spatstat generic commands as.linnet, emend, response and valid, for the class "lppm".

Value

For as.linnet.lppm a linear network (object of class "linnet"). For emend.lppm another fitted model of the same class "lppm". For response.lppm a spatial point pattern on a linear network (object of class "lpp"). For valid.lppm a logical value.

For the other methods, see the help for the default methods.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lppm, plot.lppm.

Examples

  X <- runiflpp(15, simplenet)
  fit <- lppm(X ~ x)
  print(fit)
  coef(fit)
  formula(fit)
  terms(fit)
  logLik(fit)
  deviance(fit)
  nobs(fit)
  extractAIC(fit)
  update(fit, ~1)
  valid(fit)
  vcov(fit)
  response(fit)

[Package spatstat.linnet version 3.2-5 Index]