model.frame.lppm {spatstat.linnet}	R Documentation

Extract the Variables in a Point Process Model on a Network

Description

Given a fitted point process model on a network, this function returns a data frame containing all the variables needed to fit the model using the Berman-Turner device.

Usage

 ## S3 method for class 'lppm'
model.frame(formula, ...)

Arguments

formula
A fitted point process model on a linear network. An object of class "lppm".

...
Additional arguments passed to model.frame.glm.

Details

The function model.frame is generic. This function is a method for model.frame for fitted point process models on a linear network (objects of class "lppm").

The first argument should be a fitted point process model; it has to be named formula for consistency with the generic function.

The result is a data frame containing all the variables used in fitting the model. The data frame has one row for each quadrature point used in fitting the model. The quadrature scheme can be extracted using quad.ppm.

Value

A data.frame containing all the variables used in the fitted model, plus additional variables specified in .... It has an additional attribute "terms" containing information about the model formula. For details see model.frame.glm.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A. and Turner, R. (2000) Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42, 283–322.

See Also

lppm, model.frame, model.matrix.ppm

Examples

  fit <- lppm(spiders ~ x)
  mf <- model.frame(fit)

[Package spatstat.linnet version 3.2-5 Index]