model.images.lppm {spatstat.linnet}	R Documentation

Compute Images of Constructed Covariates

Description

For a point process model fitted to spatial point pattern data on a linear network, this function computes pixel images of the covariates in the design matrix.

Usage

  ## S3 method for class 'lppm'
model.images(object, L = as.linnet(object), ...)

Arguments

object
Fitted point process model on a linear network. An object of class "lppm".

L
A linear network (object of class "linnet") in which the images should be computed. Defaults to the network in which the model was fitted.

...
Other arguments (such as na.action) passed to model.matrix.lm.

Details

This command is similar to model.matrix.lppm except that it computes pixel images of the covariates, instead of computing the covariate values at certain points only.

The object must be a fitted spatial point process model on a linear network (object of class "lppm" produced by the model-fitting function lppm).

The spatial covariates required by the model-fitting procedure are computed at every location on the network L.

Note that the spatial covariates computed here are not necessarily the original covariates that were supplied when fitting the model. Rather, they are the canonical covariates, the covariates that appear in the loglinear representation of the (conditional) intensity and in the columns of the design matrix. For example, they might include dummy or indicator variables for different levels of a factor, depending on the contrasts that are in force.

The format of the result depends on whether the original point pattern data were marked or unmarked.

If the original dataset was unmarked, the result is a named list of pixel images on the network (objects of class "linim") containing the values of the spatial covariates. The names of the list elements are the names of the covariates determined by model.matrix.lm. The result is also of class "solist" so that it can be plotted immediately.

If the original dataset was a multitype point pattern, the result is a hyperframe with one column for each possible type of points. Each column is a named list of pixel images on the network (objects of class "linim") containing the values of the spatial covariates. The row names of the hyperframe are the names of the covariates determined by model.matrix.lm.

The pixel resolution is determined by the arguments ... and spatstat.options.

Value

A list (of class "solist") or array (of class "hyperframe") containing pixel images on the network (objects of class "linim").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

model.matrix.ppm, model.matrix, lppm.

Examples

   fit <- lppm(spiders ~ x + polynom(y, 2))
   model.images(fit)

[Package spatstat.linnet version 3.2-5 Index]