model.matrix.lppm {spatstat.linnet}	R Documentation

Extract Design Matrix from Point Process Model on a Network

Description

Given a point process model that has been fitted to spatial point pattern data on a linear network, this function extracts the design matrix of the model.

Usage

   ## S3 method for class 'lppm'
model.matrix(object,
                              data=model.frame(object, na.action=NULL),
                              ..., 
                              keepNA=TRUE)

Arguments

object
The fitted point process model. An object of class "lppm".

data
A model frame, containing the data required for the Berman-Turner device.

keepNA
Logical. Determines whether rows containing NA values will be deleted or retained.

...
Other arguments (such as na.action) passed to model.matrix.lm.

Details

This is a method for the generic function model.matrix. It extracts the design matrix of a spatial point process model on a linear network (object of class "lppm").

More precisely, this command extracts the design matrix of the generalised linear model associated with a spatial point process model.

The object must be a fitted point process model on a network (object of class "lppm") produced by the model-fitting function lppm. The method model.matrix.lppm extracts the model matrix for the GLM.

The result is a matrix, with one row for every quadrature point in the fitting procedure, and one column for every canonical covariate in the design matrix.

If there are NA values in the covariates, the argument keepNA determines whether to retain or delete the corresponding rows of the model matrix. The default keepNA=TRUE is to retain them. Note that this differs from the default behaviour of many other methods for model.matrix, which typically delete rows containing NA.

Value

A matrix. Columns of the matrix are canonical covariates in the model. Rows of the matrix correspond to quadrature points in the fitting procedure (provided keepNA=TRUE).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

model.matrix, model.images.lppm, lppm

Examples

   fit <- lppm(spiders ~ x + y)
   head(model.matrix(fit))

[Package spatstat.linnet version 3.2-5 Index]