pairdist.lpp {spatstat.linnet}	R Documentation

Pairwise shortest-path distances between points on a linear network

Description

Given a pattern of points on a linear network, compute the matrix of distances between all pairs of points, measuring distance by the shortest path in the network.

Usage

## S3 method for class 'lpp'
pairdist(X, ..., method="C")

Arguments

X
Point pattern on linear network (object of class "lpp").

method
Optional string determining the method of calculation. Either "interpreted" or "C".

...
Ignored.

Details

Given a pattern of points on a linear network, this function computes the matrix of distances between all pairs of points, measuring distance by the shortest path in the network.

If two points cannot be joined by a path, the distance between them is infinite (Inf).

The argument method is not normally used. It is retained only for developers to check the validity of the software.

Value

A symmetric matrix, whose values are nonnegative numbers or infinity (Inf).

Algorithms and accuracy

Distances are accurate within the numerical tolerance of the network, summary(X)$toler.

For network data stored in the non-sparse representation described in linnet, then pairwise distances are computed using the matrix of path distances between vertices of the network, using R code if method = "interpreted", or using C code if method="C" (the default).

For networks stored in the sparse representation, the argument method has no effect, and the distances are computed using an efficient C algorithm.

Distance values

The values returned by pairdist(X) are distances, expressed as multiples of the unit of length of the spatial coordinates in X. The unit of length is given by unitname(X).

Note that, if the unit of length in X is a composite expression such as ‘2 microns’, then the values of pairdist(X) are expressed as multiples of 2 microns, rather than being expressed in microns.

Author(s)

Ang Qi Wei aqw07398@hotmail.com and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

lpp

Examples

   X <- runiflpp(12, simplenet)
   d <- pairdist(X)
   d[1:3, 1:3]

[Package spatstat.linnet version 3.2-5 Index]