pairs.linim {spatstat.linnet}	R Documentation

Scatterplot Matrix for Pixel Images on a Linear Network

Description

Produces a scatterplot matrix of the pixel values in two or more pixel images on a linear network.

Usage

## S3 method for class 'linim'
pairs(..., plot=TRUE, eps=NULL)

Arguments

...
Any number of arguments, each of which is either a pixel image on a linear network (object of class "linim"), a pixel image (object of class "im"), or a named argument to be passed to pairs.default.

plot
Logical. If TRUE, the scatterplot matrix is plotted.

eps
Optional. Spacing between sample points on the network. A positive number.

Details

This is a method for the generic function pairs for the class of pixel images on a linear network.

It produces a square array of plot panels, in which each panel shows a scatterplot of the pixel values of one image against the corresponding pixel values of another image.

At least two of the arguments ... should be a pixel image on a linear network (object of class "linim"). They should be defined on the same linear network, but may have different pixel resolutions.

First the pixel values of each image are extracted at a set of sample points equally-spaced across the network. Then pairs.default is called to plot the scatterplot matrix.

Any arguments in ... which are not pixel images will be passed to pairs.default to control the plot.

Value

Invisible. A data.frame containing the corresponding pixel values for each image. The return value also belongs to the class plotpairsim which has a plot method, so that it can be re-plotted.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

pairs.default, pairs.im

Examples

  fit <- lppm(chicago ~ marks * (x+y))
  lam <- predict(fit)
  do.call(pairs, lam)

[Package spatstat.linnet version 3.2-5 Index]