persp.linfun {spatstat.linnet}	R Documentation

Perspective View of Function on a Linear Network

Description

Given a function on a linear network, generate a perspective view.

Usage

 ## S3 method for class 'linfun'
persp(x, ..., main, eps = NULL, dimyx = NULL, xy = NULL)

Arguments

x
The function to be plotted. An object of class "linfun".

...
Arguments passed to persp.linim controlling the appearance of the plot.

main
Main title for the plot.

eps,dimyx,xy
Arguments passed to as.linim determining the spatial resolution when the function is converted to an image.

Details

The function x is converted to a pixel image on the linear network using as.linim. Then persp.linim is invoked to generate the perspective plot.

This style of plot is often attributed to Okabe and Sugihara (2012).

Value

(Invisibly) the perspective transformation matrix, as described in the help for persp.default.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Greg McSwiggan.

References

Okabe, A. and Sugihara, K. (2012) Spatial Analysis Along Networks. John Wiley and Sons, New York.

See Also

persp.linim

Examples

f <- linfun(function(x,y,seg,tp) { abs(sin(25*x)) + abs(sin(15*y)) }, simplenet)
persp(f, phi=20)

[Package spatstat.linnet version 3.2-5 Index]