plot.lpp {spatstat.linnet}	R Documentation

Plot Point Pattern on Linear Network

Description

Plots a point pattern on a linear network. Plot method for the class "lpp" of point patterns on a linear network.

Usage

## S3 method for class 'lpp'
plot(x, ..., main, add = FALSE,
     type = c("p", "n"),
     use.marks=TRUE, which.marks=NULL,
     legend=TRUE,
     leg.side=c("left", "bottom", "top", "right"),
     leg.args=list(),
     show.all = !add, show.window=FALSE, show.network=TRUE, 
     do.plot = TRUE, multiplot=TRUE) 

Arguments

x
Point pattern on a linear network (object of class "lpp").

...
Additional arguments passed to plot.linnet or plot.ppp.

main
Main title for plot.

add
Logical value indicating whether the plot is to be added to the existing plot (add=TRUE) or whether a new plot should be initialised (add=FALSE, the default).

type
Type of plot: either "p" or "n". If type="p" (the default), both the points and the observation window are plotted. If type="n", only the window is plotted.

use.marks
logical flag; if TRUE, plot points using a different plotting symbol for each mark; if FALSE, only the locations of the points will be plotted, using points().

which.marks
Index determining which column of marks to use, if the marks of x are a data frame. A character or integer vector identifying one or more columns of marks. If add=FALSE then the default is to plot all columns of marks, in a series of separate plots. If add=TRUE then only one column of marks can be plotted, and the default is which.marks=1 indicating the first column of marks.

legend
Logical value indicating whether to add a legend showing the mapping between mark values and graphical symbols (for a marked point pattern).

leg.side
Position of legend relative to main plot.

leg.args
List of additional arguments passed to plot.symbolmap or symbolmap to control the legend. In addition to arguments documented under plot.symbolmap, and graphical arguments recognised by symbolmap, the list may also include the argument sep giving the separation between the main plot and the legend, or sep.frac giving the separation as a fraction of the relevant dimension (width or height) of the main plot.

show.all
Logical value indicating whether to plot everything including the main title and the window containing the network.

show.window
Logical value indicating whether to plot the window containing the network. Overrides show.all.

show.network
Logical value indicating whether to plot the network.

do.plot
Logical value determining whether to actually perform the plotting.

multiplot
Logical value giving permission to display multiple plots.

Details

The linear network is plotted by plot.linnet, then the points are plotted using code equivalent to plot.ppp.

Commonly-used arguments include:

col and lwd for the colour and width of lines in the linear network

cols for the colour or colours of the points

chars for the plot characters representing different types of points

shape to control the shape of the symbol (this argument takes precedence over chars).

These are documented in the help file for plot.ppp.

If shape="crossticks", the points are drawn as short line segments perpendicular to the network.

Note that the linear network will be plotted even when add=TRUE, unless show.network=FALSE.

Value

(Invisible) object of class "symbolmap" giving the correspondence between mark values and plotting characters.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lpp.

See plot.ppp for options for representing the points.

See also points.lpp, text.lpp.

Examples

  plot(chicago, cols=1:7)
  plot(dendrite, shape="crossticks", cols=2:4, size=8,
                 leg.side="bottom", leg.args=list(lwd=3))

[Package spatstat.linnet version 3.2-5 Index]