plot.lppm {spatstat.linnet}	R Documentation

Plot a Fitted Point Process Model on a Linear Network

Description

Plots the fitted intensity of a point process model on a linear network.

Usage

  ## S3 method for class 'lppm'
plot(x, ..., type="trend")

Arguments

x
An object of class "lppm" representing a fitted point process model on a linear network.

...
Arguments passed to plot.linim to control the plot.

type
Character string (either "trend" or "cif") determining whether to plot the fitted first order trend or the conditional intensity.

Details

This function is the plot method for the class "lppm". It computes the fitted intensity of the point process model, and displays it using plot.linim.

The default is to display intensity values as colours. Alternatively if the argument style="width" is given, intensity values are displayed as the widths of thick lines drawn over the network.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

lppm, plot.linim, methods.lppm, predict.lppm.

Examples

  X <- runiflpp(10, simplenet)  
  fit <- lppm(X ~x)
  plot(fit)
  plot(fit, style="width")

[Package spatstat.linnet version 3.2-5 Index]