points.lpp {spatstat.linnet}	R Documentation

Draw Points on Existing Plot

Description

For a point pattern on a linear network, this function draws the coordinates of the points only, on the existing plot display.

Usage

## S3 method for class 'lpp'
points(x, ...)

Arguments

x
A point pattern on a linear network (object of class "lpp").

...
Additional arguments passed to points.default.

Details

This is a method for the generic function points for the class "lpp" of point patterns on a linear network.

If x is a point pattern on a linear network, then points(x) plots the spatial coordinates of the points only, on the existing plot display, without plotting the underlying network. It is an error to call this function if a plot has not yet been initialised.

The spatial coordinates are extracted and passed to points.default along with any extra arguments. Arguments controlling the colours and the plot symbols are interpreted by points.default. For example, if the argument col is a vector, then the ith point is drawn in the colour col[i].

Value

Null.

Difference from plot method

The more usual way to plot the points is using plot.lpp. For example plot(x) would plot both the points and the underlying network, while plot(x, add=TRUE) would plot only the points. The interpretation of arguments controlling the colours and plot symbols is different here: they determine a symbol map, as explained in the help for plot.ppp.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

plot.lpp, points.default

Examples

 plot(Frame(spiders), main="Spiders on a Brick Wall")
 points(spiders)

[Package spatstat.linnet version 3.2-5 Index]