rThomaslpp {spatstat.linnet}	R Documentation

Simulate Thomas Process on Linear Network

Description

Generate a random point pattern, a realisation of the Thomas cluster process, on a linear network.

Usage

  rThomaslpp(kappa, scale, mu, L, ..., nsim=1, drop=TRUE)

Arguments

kappa
Intensity of the Poisson process of cluster centres. A single positive number, a function(x,y), or a pixel image (object of class "im" or "linim").

scale
Standard deviation of random displacement (along the network) of a point from its cluster centre.

mu
Mean number of points per cluster (a single positive number) or reference intensity for the cluster points (a function or a pixel image).

L
Linear network (object of class "linnet") on which the point pattern should be generated.

...
Arguments passed to rpoisppOnLines.

nsim
Number of simulated realisations to generate.

drop
Logical value indicating what to do when nsim=1. If drop=TRUE (the default), the result is a point pattern. If drop=FALSE, the result is a list with one entry which is a point pattern.

Details

This function generates realisations of the Thomas cluster process on a linear network, described by Baddeley et al (2017).

Argument L can be omitted, and defaults to as.linnet(kappa), when kappa is a function on a linear network (class "linfun") or a pixel image on a linear network ("linim").

Value

A point pattern on a network (object of class "lpp") or a list of point patterns on the network.

Author(s)

Greg McSwiggan and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Nair, G., Rakshit, S. and McSwiggan, G. (2017) ‘Stationary’ point processes are uncommon on linear networks. STAT 6 (1) 68–78.

Baddeley, A., Nair, G., Rakshit, S., McSwiggan, G. and Davies, T.M. (2021) Analysing point patterns on networks — a review. Spatial Statistics 42, 100435, DOI 10.1016/j.spasta.2020.100435.

See Also

rpoislpp

Examples

  plot(rThomaslpp(4, 0.07, 5, simplenet))

[Package spatstat.linnet version 3.2-5 Index]