rcelllpp {spatstat.linnet}	R Documentation

Simulate Cell Process on Linear Network

Description

Generate a realisation of the cell process on a linear network.

Usage

rcelllpp(L, lambda, rnumgen = NULL, ..., saveid=FALSE)

Arguments

L
Either a linear network (object of class "linnet") or a tessellation on a linear network (object of class "lintess").

lambda
Intensity of the process (expected number of points per unit length),

rnumgen
Optional. Random number generator for the number of points in each cell.

...
Additional arguments to rnumgen.

saveid
Logical value indicating whether to save information about cell membership.

Details

This function generates simulated realisations of a cell point process on a network, as described in Baddeley et al (2017). This is the analogue on a linear network of the two-dimensional cell point process of Baddeley and Silverman (1988).

The argument L should be a tessellation on a linear network. Alternatively if L is a linear network, it is converted to a tessellation by treating each network segment as a tile in the tessellation.

The cell process generates a point process by generating independent point processes inside each tile of the tessellation. Within each tile, given the number of random points in the tile, the points are independent and uniformly distributed within the tile.

By default (when rnumgen is not given), the number of points in a tile of length t is a random variable with mean and variance equal to lambda * t, generated by calling rcellnumber.

If rnumgen is given, it should be a function with arguments rnumgen(n, mu, ...) where n is the number of random integers to be generated, mu is the mean value of the distribution, and ... are additional arguments, if needed. It will be called in the form rnumgen(1, lambda * t, ...) to determine the number of random points falling in each tile of length t.

Value

Point pattern on a linear network (object of class "lpp"). If saveid=TRUE, the result has an attribute "cellid" which is a factor specifying the cell that contains each point.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A.J. and Silverman, B.W. (1984) A cautionary example on the use of second-order methods for analyzing point patterns. Biometrics 40, 1089-1094.

Baddeley, A., Nair, G., Rakshit, S. and McSwiggan, G. (2017) ‘Stationary’ point processes are uncommon on linear networks. STAT 6, 68–78.

See Also

rSwitzerlpp

Examples

   X <- rcelllpp(domain(spiders), 0.01)
   plot(X)
   plot(linearK(X))

[Package spatstat.linnet version 3.2-5 Index]