repairNetwork {spatstat.linnet}	R Documentation

Repair Internal Data in a Linear Network

Description

Detect and repair inconsistencies or duplication in the internal data of a network object.

Usage

repairNetwork(X)

Arguments

X
A linear network (object of class "linnet") or a point pattern on a linear network (object of class "lpp").

Details

This function detects and repairs inconsistencies in the internal data of X. Currently it does the following:

checks that different ways of calculating the number of edges give the same answer

removes any duplicated edges of the network

ensures that each edge is recorded as a pair of vertex indices (from, to) with from < to.

Value

An object of the same kind as X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

thinNetwork

[Package spatstat.linnet version 3.2-5 Index]