sdr.lpp {spatstat.linnet}	R Documentation

Sufficient Dimension Reduction for a Point Pattern on a Linear Network

Description

Given a point pattern on a linear network, and a set of predictors, find a minimal set of new predictors, each constructed as a linear combination of the original predictors.

Usage

## S3 method for class 'lpp'
sdr(X, covariates,
            method = c("DR", "NNIR", "SAVE", "SIR", "TSE"),
            Dim1 = 1, Dim2 = 1, predict=FALSE, ...)

Arguments

X
A point pattern on a linear network (object of class "lpp").

covariates
A list of pixel images (objects of class "im" or "linim") to serve as predictor variables.

method
Character string indicating which method to use. See Details.

Dim1
Dimension of the first order Central Intensity Subspace (applicable when method is "DR", "NNIR", "SAVE" or "TSE").

Dim2
Dimension of the second order Central Intensity Subspace (applicable when method="TSE").

predict
Logical value indicating whether to compute the new predictors as well.

...
Extra arguments are ignored.

Details

This is the method for sdr for the class "lpp" of point patterns on a linear network.

Given a point pattern X and predictor variables Z[1], ..., Z[p], Sufficient Dimension Reduction methods (Guan and Wang, 2010) attempt to find a minimal set of new predictor variables, each constructed by taking a linear combination of the original predictors, which explain the dependence of X on Z[1], ..., Z[p]. The methods do not assume any particular form of dependence of the point pattern on the predictors. The predictors are assumed to be Gaussian random fields.

Available methods are:

method="DR" 	 directional regression

method="NNIR" 	 nearest neighbour inverse regression

method="SAVE" & sliced average variance estimation

method="SIR" & sliced inverse regression

method="TSE" & two-step estimation

The result includes a matrix B whose columns are estimates of the basis vectors of the space of new predictors. That is, the jth column of B expresses the jth new predictor as a linear combination of the original predictors.

If predict=TRUE, the new predictors are also evaluated. They can also be evaluated using sdrPredict.

Value

A list with components B, M or B, M1, M2 where B is a matrix whose columns are estimates of the basis vectors for the space, and M or M1,M2 are matrices containing estimates of the kernel.

If predict=TRUE, the result also includes a component Y which is a list of pixel images giving the values of the new predictors.

Author(s)

Based on a Matlab original, for two-dimensional point patterns, by Yongtao Guan. Adapted to R, and to linear networks, by Suman Rakshit.

References

Guan, Y. and Wang, H. (2010) Sufficient dimension reduction for spatial point processes directed by Gaussian random fields. Journal of the Royal Statistical Society, Series B, 72, 367–387.

See Also

sdrPredict to compute the new predictors from the coefficient matrix.

dimhat to estimate the subspace dimension.

subspaceDistance

Examples

#   sdr(bei, bei.extra)

   xim <- as.linim(function(x,y) { x }, simplenet)
   yim <- as.linim(function(x,y) { y }, simplenet)
   X <- runiflpp(30, simplenet)
   sdr(X, list(x=xim, y=yim))

[Package spatstat.linnet version 3.2-5 Index]