spatstat.linnet-deprecated {spatstat.linnet}	R Documentation

Deprecated spatstat.linnet functions

Description

Deprecated spatstat.linnet functions.

Usage

## S3 method for class 'linnet'
circumradius(x, ...)
rjitterlpp(X, ...)
bw.relrisklpp(X, ...)

Details

These functions are deprecated, and will eventually be deleted from the spatstat.linnet package.

rjitterlpp is replaced by rjitter.lpp, a method for the generic rjitter.

circumradius.linnet is replaced by the more appropriately named boundingradius.linnet.

bw.relrisklpp is replaced by the method bw.relrisk.lpp for the generic bw.relrisk.

Value

circumradius.linnet returns a numeric value.

rjitterlpp returns a point pattern on a linear network (object of class "lpp").

bw.relrisklpp returns an object of class "bw.optim" which is a numeric value with additional attributes.

[Package spatstat.linnet version 3.2-5 Index]