superimpose.lpp {spatstat.linnet}	R Documentation

Superimpose Several Point Patterns on Linear Network

Description

Superimpose any number of point patterns on the same linear network.

Usage

  ## S3 method for class 'lpp'
superimpose(..., L=NULL)

Arguments

...
Any number of arguments, each of which represents a point pattern on the same linear network. Each argument can be either an object of class "lpp", giving both the spatial coordinates of the points and the linear network, or a list(x,y) or list(x,y,seg,tp) giving just the spatial coordinates of the points.

L
Optional. The linear network. An object of class "linnet". This argument is required if none of the other arguments is of class "lpp".

Details

This function is used to superimpose several point patterns on the same linear network. It is a method for the generic function superimpose.

Each of the arguments ... can be either a point pattern on a linear network (object of class "lpp" giving both the spatial coordinates of the points and the linear network), or a list(x,y) or list(x,y,seg,tp) giving just the spatial coordinates of the points. These arguments must represent point patterns on the same linear network.

The argument L is an alternative way to specify the linear network, and is required if none of the arguments ... is an object of class "lpp".

The arguments ... may be marked patterns. The marks of each component pattern must have the same format. Numeric and character marks may be “mixed”. If there is such mixing then the numeric marks are coerced to character in the combining process. If the mark structures are all data frames, then these data frames must have the same number of columns and identical column names.

If the arguments ... are given in the form name=value, then the names will be used as an extra column of marks attached to the elements of the corresponding patterns.

Value

An object of class "lpp" representing the combined point pattern on the linear network.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

Ege Rubak rubak@math.aau.dk

and Greg McSwiggan.

See Also

superimpose

Examples

  X <- rpoislpp(5, simplenet)
  Y <- rpoislpp(10, simplenet)
  superimpose(X,Y) # not marked
  superimpose(A=X, B=Y) # multitype with types A and B

[Package spatstat.linnet version 3.2-5 Index]