terminalvertices {spatstat.linnet}	R Documentation

Terminal Vertices of a Linear Network

Description

Finds the terminal vertices of a linear network.

Usage

 terminalvertices(L)

Arguments

L
A linear network (object of class "linnet").

Details

Given the linear network L, this function examines the vertices (segment endpoints) of the network and determines which of them are ‘terminal’ vertices (i.e. the endpoint of only one segment). These terminal vertices are returned as a point pattern on the network.

Value

A point pattern on the same linear network (object of class "lpp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk and Mehdi Moradi m2.moradi@yahoo.com.

See Also

vertices.linnet.

Examples

  B <- terminalvertices(simplenet)
  plot(simplenet, main="")
  plot(B, add=TRUE, pch=16, cex=2)

[Package spatstat.linnet version 3.2-5 Index]