tile.lengths {spatstat.linnet}	R Documentation

Compute Lengths of Tiles in a Tessellation on a Network

Description

Computes the length of each tile in a tessellation on a linear network.

Usage

tile.lengths(x)

Arguments

x
A tessellation on a linear network (object of class "lintess").

Details

A tessellation on a linear network L is a partition of the network into non-overlapping pieces (tiles). Each tile consists of one or more line segments which are subsets of the line segments making up the network. A tile can consist of several disjoint pieces.

This command computes the length of each of the tiles that make up the tessellation x. The result is a numeric vector.

Value

A numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lintess

Examples

  X <- runiflpp(5, simplenet)
  A <- lineardirichlet(X)
  plot(A)
  tile.lengths(A)

[Package spatstat.linnet version 3.2-5 Index]