is.connected.linnet {spatstat.linnet}	R Documentation

Determine Whether a Linear Network is Connected

Description

Determine whether a linear network is topologically connected.

Usage

## S3 method for class 'linnet'
is.connected(X, ...)

Arguments

X
A linear network (object of class "linnet").

...
Arguments passed to connected.linnet to determine the connected components.

Details

The command is.connected(X) returns TRUE if the network X consists of a single, topologically-connected piece, and returns FALSE if X consists of several pieces which are not joined together.

The function is.connected is generic, with methods for several classes. This help file documents the method for linear networks, is.connected.linnet.

Value

A logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

is.connected, connected, connected.lpp.

Examples

  is.connected(simplenet)

[Package spatstat.linnet version 3.2-5 Index]