is.multitype.lpp {spatstat.linnet}	R Documentation

Test Whether A Point Pattern on a Network is Multitype

Description

Tests whether a point pattern on a network has “marks” attached to the points which classify the points into several types.

Usage

  ## S3 method for class 'lpp'
is.multitype(X, na.action="warn", ...) 

Arguments

X
Point pattern on a linear networl (object of class "lpp").

na.action
String indicating what to do if NA values are encountered amongst the marks. Options are "warn", "fatal" and "ignore".

...
Ignored.

Details

“Marks” are observations attached to each point of a point pattern. For example the chicago dataset contains the locations of crimes, each crime location being marked by the type of crime.

This function tests whether the point pattern X contains or involves marked points, and that the marks are a factor. It is a method for the generic function is.multitype.

The argument na.action determines what action will be taken if the point pattern has a vector of marks but some or all of the marks are NA. Options are "fatal" to cause a fatal error; "warn" to issue a warning and then return TRUE; and "ignore" to take no action except returning TRUE.

Value

Logical value, equal to TRUE if X is a multitype point pattern.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

is.multitype, is.multitype.lppm

Examples

   is.multitype(chicago)

[Package spatstat.linnet version 3.2-5 Index]