linearJinhom {spatstat.linnet}	R Documentation

Inhomogeneous Linear J-function for Point Processes on Linear Networks

Description

Computes an estimate of the inhomogeneous linear J-function for a point pattern on a linear network.

Usage

linearJinhom(X, lambda = NULL, lmin=NULL,
             ...,
             r=NULL, rmax=NULL,
             distance=c("path","euclidean"),
             densitymethod=c("kernel", "Voronoi"),
             sigma=bw.scott.iso,
             f=0.2, nrep=200, ngrid=256) 

Arguments

X
Point pattern on linear network (object of class "lpp").

lambda
Intensity values for the point pattern. Either a numeric vector, a function, a pixel image (object of class "im" or "linim") or a fitted point process model (object of class "ppm" or "lppm").

lmin
Optional. The minimum possible value of the intensity over the network. A positive numerical value.

r
Optional. Numeric vector of values of the function argument r. There is a sensible default.

rmax
Optional. Numeric value specifying the largest desired value of r. There is a sensible default.

distance
A string (partially matched) specifying the metric that will be used to measure distances between points on the network: distance="path" is the shortest-path distance, and distance="euclidean" is the Euclidean distance.

densitymethod
String (partially matched) specifying the method that will be used to estimate the intensity lambda, if lambda is not given: densitymethod="kernel" specifies kernel smoothing and densitymethod="Voronoi" specifies Voronoi estimation. See Details.

sigma
Smoothing bandwidth used to estimate lambda by kernel smoothing, if lambda is not given and densitymethod="kernel". Either a numeric value, or a function that can be applied to X to compute the bandwidth.

f,nrep
Arguments passed to the algorithm for estimating the intensity by Voronoi estimation, if lambda is not given and densitymethod="Voronoi".

...
Additional arguments passed to the algorithms that estimate the intensity, if lambda is not given.

ngrid
Integer specifying the number of sample points on the network that will be used to estimate the inhomogeneous empty space function F.

Details

This function computes the geometrically corrected inhomogeneous linear J-function for point processes on linear networks defined by Cronie et al (2020).

The argument lambda is the (estimated) intensity of the underlying point process. It should be either a numeric vector (giving intensity values at the points of X), a function, a pixel image (object of class "im" or "linim") or a fitted point process model (object of class "ppm" or "lppm").

If lambda is not given, it will be estimated from the observed point pattern X as follows:

If densitymethod="kernel", the intensity will be estimated by kernel smoothing, using the fast estimator densityQuick.lpp introduced by Rakshit et al (2019). The smoothing bandwidth sigma is required. It may be specified as a numeric value, or as a function that can be applied to X to obtain a bandwidth value. Examples of the latter include bw.scott.iso and bw.lppl. Additional arguments ... will be passed to sigma and to densityQuick.lpp.

If densitymethod = "Voronoi", the intensity will be estimated using the resample-smoothed Voronoi estimator densityVoronoi.lpp introduced by Moradi et al (2019). The arguments f and nrep are passed to densityVoronoi.lpp and determine the retention probability and the number of replicates, respectively. Additional arguments ... will be passed to densityVoronoi.lpp.

Value

Function value table (object of class "fv").

Author(s)

Mehdi Moradi m2.moradi@yahoo.com and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Cronie, O., Moradi, M., and Mateu, J. (2020) Inhomogeneous higher-order summary statistics for point processes on linear networks. Statistics and Computing 30 (6) 1221–1239.

Moradi, M., Cronie, 0., Rubak, E., Lachieze-Rey, R., Mateu, J. and Baddeley, A. (2019) Resample-smoothing of Voronoi intensity estimators. Statistics and Computing 29 (5) 995–1010.

Rakshit, S., Davies, T., Moradi, M., McSwiggan, G., Nair, G., Mateu, J. and Baddeley, A. (2019) Fast kernel smoothing of point patterns on a large network using 2D convolution. International Statistical Review 87 (3) 531–556. DOI: 10.1111/insr.12327.

See Also

bw.scott.iso, bw.lppl, densityVoronoi.lpp, densityQuick.lpp

linearKinhom

Jinhom

Examples

  if(interactive()) {
    plot(linearJinhom(spiders))
  } else {
    bottomhalf <- owin(c(0, 1125), c(0, 500))
    plot(linearJinhom(spiders[bottomhalf]))
  }

[Package spatstat.linnet version 3.2-5 Index]