linearpcfinhom {spatstat.linnet}	R Documentation

Inhomogeneous Linear Pair Correlation Function

Description

Computes an estimate of the inhomogeneous linear pair correlation function for a point pattern on a linear network.

Usage

linearpcfinhom(X, lambda=NULL, r=NULL, ..., correction="Ang",
               normalise=TRUE, normpower=1,
	       update = TRUE, leaveoneout = TRUE,
               sigma=NULL, adjust.sigma=1,
               bw="nrd0", adjust.bw=1,
	       ratio = FALSE)

Arguments

X
Point pattern on linear network (object of class "lpp").

lambda
Intensity values for the point pattern. Either a numeric vector, a function, a pixel image (object of class "im") or a fitted point process model (object of class "ppm" or "lppm").

r
Optional. Numeric vector of values of the function argument r. There is a sensible default.

...
Arguments passed to density.default to control the smoothing of the estimates of pair correlation.

correction
Geometry correction. Either "none" or "Ang". See Details.

normalise
Logical. If TRUE (the default), the denominator of the estimator is data-dependent (equal to the sum of the reciprocal intensities at the data points, raised to normpower), which reduces the sampling variability. If FALSE, the denominator is the length of the network.

normpower
Integer (usually either 1 or 2). Normalisation power. See explanation in linearKinhom.

update
Logical value indicating what to do when lambda is a fitted model (class "lppm" or "ppm"). If update=TRUE (the default), the model will first be refitted to the data X (using update.lppm or update.ppm) before the fitted intensity is computed. If update=FALSE, the fitted intensity of the model will be computed without re-fitting it to X.

leaveoneout
Logical value specifying whether to use a leave-one-out rule when calculating the intensity. See Details.

sigma
Smoothing bandwidth (passed to density.lpp) for kernel density estimation of the intensity when lambda=NULL.

adjust.sigma
Numeric value. sigma will be multiplied by this value.

bw
Smoothing bandwidth (passed to density.default) for one-dimensional kernel smoothing of the pair correlation function. Either a numeric value, or a character string recognised by density.default.

adjust.bw
Numeric value. bw will be multiplied by this value.

ratio
Logical. If TRUE, the numerator and denominator of each estimate will also be saved, for use in analysing replicated point patterns.

Details

This command computes the inhomogeneous version of the linear pair correlation function from point pattern data on a linear network.

The argument lambda should provide estimated values of the intensity of the point process at each point of X.

If lambda=NULL, the intensity will be estimated by kernel smoothing by calling density.lpp with the smoothing bandwidth sigma, and with any other relevant arguments that might be present in .... A leave-one-out kernel estimate will be computed if leaveoneout=TRUE.

If lambda is given, it may be a numeric vector (of length equal to the number of points in X), or a function(x,y) that will be evaluated at the points of X to yield numeric values, or a pixel image (object of class "im") or a fitted point process model (object of class "ppm" or "lppm").

If lambda is a fitted point process model, the default behaviour is to update the model by re-fitting it to the data, before computing the fitted intensity. This can be disabled by setting update=FALSE. The intensity at data points will be computed by fitted.lppm or fitted.ppm. A leave-one-out estimate will be computed if leaveoneout=TRUE and update=TRUE.

If correction="none", the calculations do not include any correction for the geometry of the linear network. If correction="Ang", the pair counts are weighted using Ang's correction (Ang, 2010).

The bandwidth for smoothing the pairwise distances is determined by arguments ... passed to density.default, mainly the arguments bw and adjust. The default is to choose the bandwidth by Silverman's rule of thumb bw="nrd0" explained in density.default.

Value

Function value table (object of class "fv").

If ratio=TRUE then the return value also has two attributes called "numerator" and "denominator" which are "fv" objects containing the numerators and denominators of each estimate of g(r).

Warning

Older versions of linearpcfinhom interpreted lambda=NULL to mean that the homogeneous function linearpcf should be computed. This was changed to the current behaviour in version 3.1-0 of spatstat.linnet.

Author(s)

Ang Qi Wei aqw07398@hotmail.com and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Ang, Q.W. (2010) Statistical methodology for spatial point patterns on a linear network. MSc thesis, University of Western Australia.

Ang, Q.W., Baddeley, A. and Nair, G. (2012) Geometrically corrected second-order analysis of events on a linear network, with applications to ecology and criminology. Scandinavian Journal of Statistics 39, 591–617.

Okabe, A. and Yamada, I. (2001) The K-function method on a network and its computational implementation. Geographical Analysis 33, 271-290.

See Also

linearpcf, linearKinhom, lpp

Examples

  X <- rpoislpp(5, simplenet)
  fit <- lppm(X ~x)
  g <- linearpcfinhom(X, lambda=fit, update=FALSE)
  plot(g)
  ge <- linearpcfinhom(X, sigma=bw.lppl)

[Package spatstat.linnet version 3.2-5 Index]