nnfromvertex {spatstat.linnet}	R Documentation

Nearest Data Point From Each Vertex in a Network

Description

Given a point pattern on a linear network, for each vertex of the network find the nearest data point.

Usage

nnfromvertex(X, what = c("dist", "which"), k = 1)

Arguments

X
Point pattern on a linear network (object of class "lpp").

what
Character string specifying whether to return the nearest-neighbour distances, nearest-neighbour identifiers, or both.

k
Integer, or integer vector, specifying that the kth nearest neighbour should be returned.

Details

For each vertex (node) of the linear network, this algorithm finds the nearest data point to the vertex, and returns either the distance from the vertex to its nearest neighbour in X, or the serial number of the nearest neighbour in X, or both.

If k is an integer, then the k-th nearest neighbour is found instead.

If k is an integer vector, this is repeated for each integer in k.

Value

A numeric vector, matrix, or data frame.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

nndist.lpp

Examples

  X <- runiflpp(5, simplenet)
  nnfromvertex(X)
  nnfromvertex(X, k=1:3)

[Package spatstat.linnet version 3.2-5 Index]