simulate.lppm {spatstat.linnet}	R Documentation

Simulate a Fitted Point Process Model on a Linear Network

Description

Generates simulated realisations from a fitted Poisson point process model on a linear network.

Usage

  ## S3 method for class 'lppm'
simulate(object, nsim=1, ...,
                         new.coef=NULL,
                         progress=(nsim > 1),
                         drop=FALSE)

Arguments

object
Fitted point process model on a linear network. An object of class "lppm".

nsim
Number of simulated realisations.

progress
Logical flag indicating whether to print progress reports for the sequence of simulations.

new.coef
New values for the canonical parameters of the model. A numeric vector of the same length as coef(object).

...
Arguments passed to predict.lppm to determine the spatial resolution of the image of the fitted intensity used in the simulation.

drop
Logical. If nsim=1 and drop=TRUE, the result will be a point pattern, rather than a list containing a point pattern.

Details

This function is a method for the generic function simulate for the class "lppm" of fitted point process models on a linear network.

Only Poisson process models are supported so far.

Simulations are performed by rpoislpp.

Value

A list of length nsim containing simulated point patterns (objects of class "lpp") on the same linear network as the original data used to fit the model. The result also belongs to the class "solist", so that it can be plotted, and the class "timed", so that the total computation time is recorded.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

, Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

lppm, rpoislpp, simulate

Examples

  fit <- lppm(unmark(chicago) ~ y)
  simulate(fit)[[1]]

[Package spatstat.linnet version 3.2-5 Index]