tilenames.lintess {spatstat.linnet}	R Documentation

Names of Tiles in a Tessellation on a Network

Description

Extract or Change the Names of the Tiles in a Tessellation on a Network.

Usage

## S3 method for class 'lintess'
tilenames(x)

## S3 replacement method for class 'lintess'
tilenames(x) <- value

Arguments

x
A tessellation on a linear network (object of class "lintess").

value
Character vector giving new names for the tiles.

Details

These functions extract or change the names of the tiles that make up the tessellation x.

If the tessellation is a regular grid, the tile names cannot be changed.

Value

tilenames returns a character vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

lintess, tiles

Examples

  B <- lineardirichlet(runiflpp(5, simplenet))
  tilenames(B)
  tilenames(B) <- letters[1:5]

[Package spatstat.linnet version 3.2-5 Index]